requirements.txt
attrs==24.2.0
autopep8==2.3.1
black==24.10.0
blinker==1.8.2
click==8.1.7
distlib==0.3.9
dnspython==2.7.0
filelock==3.16.1
flake8==7.1.1
flasgger==0.9.7.1
Flask==3.0.3
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.1
mccabe==0.7.0
mistune==3.0.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pbr==6.1.0
platformdirs==4.3.6
print-color==0.4.6
pycodestyle==2.12.1
pyflakes==3.2.0
pymongo==4.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
stevedore==5.3.0
virtualenv==20.27.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
Werkzeug==3.0.4
